Projectiles will be analized in EntityDamageByEntityEvent
Now the plugin will be future proof to any new type of projectile added in the game. Now players wont be damaged if they are being hit by any projectile shot by another player. This replaces the Arrow and Trident check as both extend from Projectile class. This also considers InstantDamage potions but not other types as EntityDamageByEntityEvent is not being called (though there is another system for preventint potion damage so wont be a problem)