Fixed an issue with /ps info
Basicly when you did /ps info without admin perms in a region which only has one field the command would tell you the owner. If you did in one which had many fields it wouldnt send you a message. I basicly made it to tell you the owner of the latest field that it checked (idk if you want to give a list of owners or whatever, just a quick fix)